+++ src/translations/nl/freeform.php
@@ -547,6 +547,8 @@ return [
 
     'Enable AJAX'                                             => 'Schakel AJAX in',
     'Use Freeform\'s built-in automatic AJAX submit feature.' => 'Geburik Freeform\'s ingebouwde automatische AJAX verzendfunctie.',
+    'Enable reCAPTCHA' => 'Schakel reCAPTCHA in',
+    'Disabling this option removes the reCAPTCHA check for this specific form.' => 'Als u deze optie uitschakelt, wordt de reCAPTCHA-controle voor dit specifieke formulier verwijderd.',
 
     'Input attributes'                            => 'Input attributen',
     'Label attributes'                            => 'Label attributen',